requirements.txt
Flask==2.3.3
click==8.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
Werkzeug==2.3.7
importlib-metadata==3.6.0
blinker==1.6.2
MarkupSafe==2.1.1
zipp==3.1.0
Flask-HTTPAuth==4.8.0
Flask-Cors==4.0.0
flask-openapi3==2.5.2
pydantic==1.10.12
typing-extensions==4.8.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
Flask-SQLAlchemy==3.1.1
aniso8601==9.0.1
attrs==23.1.0
PyJWT==2.8.0
flask-restx==1.1.0
importlib-resources==6.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
pkgutil-resolve-name==1.3.10
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.3
python-dotenv==1.0.0